requirements.txt
numpy
scipy
matplotlib
python-dotenv
plotly
